Extract path optimization
