+++ docs/refactor-architecture-plan.md
@@ -639,6 +639,7 @@ Use this section to track completed steps with date and PR/commit references.
 - 2026-04-10: Full-suite validation confirmed by user after FluidNC modularization slices; full suite passing.
 - 2026-04-10: Added pre-refactor gcodeEngine behavior hardening before continuing pipeline extraction: ran focused baseline coverage with `npx.cmd vitest run tests/unit/gcodeEngine.test.ts --coverage --coverage.include=src/workers/gcodeEngine.ts` (77/77) showing `gcodeEngine.ts` at 97.61% statements, 91.39% branches, 100% functions, 99.01% lines with the nearest-neighbour fallback branch uncovered; added a deterministic edge-case test in `tests/unit/gcodeEngine.test.ts` (`falls back to linear scan when distance comparisons are non-finite`) to lock existing fallback behavior; revalidated with `npm.cmd run test -- tests/unit/gcodeEngine.test.ts` (78/78) and updated focused coverage now at 98.7% statements, 93.01% branches, 100% functions, 100% lines.
 - 2026-04-10: Continued gcodeEngine pipeline modularization with stage extraction of coordinate transform + curve flattening: moved `transformPt`, `cubicBezier`, `quadBezier`, and `arcToBeziers` into `src/workers/gcodeEngine/stages/geometryFlattening.ts`, kept backward-compatible re-exports in `src/workers/gcodeEngine.ts`, and added focused direct stage coverage in `tests/unit/gcodeGeometryFlattening.test.ts`.
+- 2026-04-10: Continued gcodeEngine pipeline modularization with path optimization stage extraction: moved `nearestNeighbourSort` + `joinSubpaths` into `src/workers/gcodeEngine/stages/pathOptimization.ts`, kept backward-compatible re-exports in `src/workers/gcodeEngine.ts`, and added focused deterministic stage coverage in `tests/unit/gcodePathOptimization.test.ts`.
 - 2026-04-10: Started gcodeEngine pipeline modularization with stage extraction of path parsing: moved `tokenizePath` + `toAbsolute` (and `PathToken`) into `src/workers/gcodeEngine/stages/pathParsing.ts`, kept backward-compatible exports in `src/workers/gcodeEngine.ts`, and added focused deterministic coverage in `tests/unit/gcodePathParsing.test.ts`.
 
 ## Update Rule
